Installer package and DMG creation automation.
create-dmg.sh now creates a DMG with an Installer package.
The install doesn't yet create a working yeti install because
of library hell. Also, the node binary needs to be fattened up
for i386 in addition to x86_64 archs.

Requires the Iceberg package creator, which is much better than
trying to get PackageMaker to work on the command line.