Prevent race when incoming() is called rapidly.
Calling YETI.next() when we already are processing
tests will cause the current test to get bumped.
This is a stopgap fix to prevent this.

This really only happens in our unit tests because
we test seperate batches of fixture tests in rapid
succession, which rarely happens in actual use.