Graceful shutdown on SIGINT.
When Ctrl-C is used to shutdown the Yeti server,
connected browsers are told to close their connections
and don't reconnect.

This should prevent interference during unit tests
from browsers trying to reconnect on the same port
that were launched during previous runs.

Safari seems to have an EventSource related bug that
may manifest itself when it tries too many times to
reconnect to a dead EventSource. I hope this helps.