Bugfixes: Clone the urlQueue, Yoshi workaround.
EventYoshi appears to invoke all listeners for every listener
attached to the EventYoshi for that event. This bug is rarely
triggered because we rarely listen to the same event twice.
Workaround by listening to our own event instead.

Remove duplicate agentComplete emit.

Invoke agent.dispatch with a copy of the urlQueue instead of
the reference to the urlQueue.