Inject script near the doctype. Ref #64.
Without using a DOM, it's difficult to determine where
the first instance of "<script" is that isn't inside
an HTML comment, stylesheet, iframe, etc. -- there
are a lot of tags that have to be checked in the markup
preceding a match of "<script". It's possible, but
that would be fragile.

Instead, attempt to inject into after the <head>, <html>,
or <!doctype html> tags if present.

WARNING: This commit does not yet check for <script>
in between a match: for example, "<html><script/><head>"
would not work as expected. Future commits will verify
that only whitespace or another tag token (i.e. "<head>"
or "<html") exists between the match and beginning of
the file.