Fix #63, do not skip tests with 404ing resources.
All resources, such as scripts linked from a test file,
used to recieve a 302 Found response with the next test URL
in the Location header if the resource didn't exist.

This is bad for many reasons, including that calling nextURL
and attempting to serve a 302 for a script or image would
cause a test to become skipped.

Only serve 302s for HTML files that are a part of the batch.
Record an error for other 404s and serve the correct 404 code.