Added `onbeforeunload` to track browser exits
The `onbeforeunload` event will trigger a
dynamic image that loads a url like:
`/ping/unload/AGENT_ID`

This will allow the server to set the agent
to a disconnected state so it won't be processed
any longer.

This is safer than a socket connection or an
XHR request since it's a simple ping with a
dead image.