Prevent race during IE6 navigation delay.
Additionally replace use of beforeunload event with unload.

If a reconnection occurs after IE6 page navigation is
scheduled, the moving bit would be reset which would trigger
an unload ping when the page moves to the next test.

Instead of using flags, attach the unload handler using
browser event APIs and detach the handler before a controlled
move to the next test.