Provide query CLI option to the server.
The query option specifies a query-string parameter
that will be added during testing. This is used
to trigger YUI Loader to load instrumented files
for code coverage, e.g. with filter=coverage.

Note: router support for the query string does not exist yet,
so files requested will contain the query part of the URL
as the filename.