Refactor Batch-related APIs to use options object.
Instead of changing several argument lists when adding
a new parameter to the Batch, use an object with expandable
properties. These properties are defined in YUIDoc code
comments above `Client.createBatch()`.

Introduce `util.shallowClone()` on the new `util` module,
which is an extension of the built-in `util`. When modifying
the options object argument, copy the object first using
`shallowClone` to prevent breaking API users, including
our own tests.