YUI Test Coverage support. Display line coverage.
If tested code was instrumented with YUI Test Coverage,
coverage data will be passed along with test results
in the agentResults event.

The `./go yui` script uses the recently-added `--query`
command-line option to run YUI's tests with `filter=coverage`
in the query string, which triggers YUI Loader to load
instrumented files.

A basic line coverage percentage is displayed during testing
on the command line when coverage data is available.