Reset heartbeat timer when test files are served.
The heartbeat timer is only reset when YUI Test reports
a test pass or fail event. If a test fails to run any YUI Tests
before the heartbeat timeout of 45 seconds, then future tests
will immediately hit this timeout because the timeout is not
cleared when a new test is requested.

This fix causes the timer to reset when test-related files
are served, which includes requesting a new test.

Credit to John Traver <jtraver@yahoo-inc.com> for
reporting this issue.