Set timeouts per batch, enforce timeouts.
Accept a `timeout` option in createBatch
and a `--timeout` option on the command line
to specify an agent timeout in seconds.

No longer use a `waiting` flag for agents on the
capture page. Instead, the capture page now checks
in with the Hub every 10 seconds which resets the
server-side agent timeout.

Change the server-side agent timeout to 45 seconds,
up from 3.6 seconds.