AgentManager reschedules timer when TTL < 10 sec.
The AgentManager sets an interval for agent reaping
every 10 seconds. If agent timeouts are < 10 seconds,
then the timeout may be delayed as much as 10 seconds
before the reap check happens.

Reschedule the AgentManager's reap timer when an Agent
requests a TTL under 10 seconds.

Add a functional test for timeout behavior.