Configurable CLI basedir.
Let a .yeti.json define the basedir with a relative path
to be resolved from the path where the .yeti.json is located.

For example, a .yeti.json inside /foo/bar setting basedir to `.`
would cause the CLI running inside /foo/bar/baz/quux to use
/foo/bar as the basedir, because `.` resolves to the directory
where the configuration was stored.

Files submitted to Yeti's Batch API need to have paths relative
to the basedir, so resolve files to the basedir before sending.

For example, the CLI running in /foo/bar/baz/quux for test.html
would need to specify baz/quux/test.html as the file to the Yeti
Batch API, relative to the given basedir /foo/bar.