Override .yeti.json hub option with `--no-hub`.
Change nopt hub option type to [url, Boolean].

If the CLI provided hub value is not a valid URL,
the value will be `true`. If that happens, print
a stderr message and discard the bad value.

If the CLI provided hub value is `false`, that
means `--no-hub` or `--hub false` was provided.
Let that value be imported into the configuration
so the CLI will know to start a new Hub.