CLI stdin/out/err streams; evented mocks. Ref #95.
The CLI now expects to be provided with streams for:

 - stdin
 - stdout
 - stderr

This replaces the old options:

 - readableStream
 - writableStream
 - putsFn
 - errorFn

Removing putsFn and errorFn discourages usage of Node.js
stdio (e.g., console.*), which is one less thing to mock.
The CLI.prototype's puts() and error() methods still allow
for `printf`-like formatting by using util.format().

Create new mock Readable and Writable Streams for testing.

The MockWritableStream supports expect(), which will fire
a callback when the expected string is provided. This lets
us do async testing of the CLI.