Distribute batch workload in parallel. Fix #109.
Allows for multiple instances of the same browser
to distribute the batch workload for that browser.

The workload unit is a HTML test file, so batches
need to be leverage multiple files for this to be
beneficial.

Before this change, Batch objects would request
to use all available Agent objects when dispatching.

Now, Batch objects now request AgentGroup objects.
An AgentGroup contains one or more Agent objects
and is responsible for delegating batch workload
among the agents under its control.

AgentGroup objects are created on batch creation.
In the future, AgentGroup objects may be created
for other criteria besides a User-Agent match.

The AgentManager mainly deals with Agent objects
instead of AgentGroups. When determining if an
Agent is groupable, it considers if it already
belongs to a group. In the future, the AgentManager
may choose to use a subset of available agents
to create groups.