Fix test: Local URL does not appear when offline.
The "also available locally" text with the local network IP
does not appear when the computer is not connected to any
network -- like when you have Wi-Fi turned off, for example.

This causes false failures when offline, so remove the check.
The local IP display is supposed to be optional.