Do now allow >1 socket with the same Agent ID.
This prevents cut-and-paste Yeti capture page URLs from
causing two different sockets to become connected to
the same Agent socketEmitter.

Agent objects should never have more than one socketEmitter
child. If that happens, do not add the socket to the Agent
and tell the Agent to request a new ID from the server.