Add ci-coverage.sh for Jenkins CI builds.
Istanbul at master is used to make this work.
We use Istanbul and Vows at the same time.
Vows prints test results to stdout, but even
with the `--print none` option, Istanbul will
print some informational messages to stdout.

These messages are printed to stderr on master,
so the XML on stdout doesn't have these messages.