Do not resend results when we will navigate soon.
Emitting results again after we have been given an order
to navigate will cause the server to emit another navigate event
and will reset the navigateTimeout. The ping handler will
resendResultsIfNeeded -- if that happens with an active navigateTimeout
because the navigationTimeoutMilliseconds value is high (e.g., IE 6,7)
then a loop is created as the browser responds to pings.