Revert use of request because of yogi conflict.
Yeti is used in the Yogi tool which uses request.
During Yeti's postinstall in Yogi's installation,
Yogi's request is used before all of request's
dependencies are fully installed.

This is probably a bug in npm that needs to
be resolved. But for now, we need to back out
our use of request in our postinstall to unblock
Yogi users.

This will break the use of HTTP_PROXY env vars
introduced in GH-34.

This is a squash of two reverts listed below.

Revert "Fix saveURLToDep to properly write files to disk."

This reverts commit c6474090aee4e0cf9e87652a0166eaab5ddac325.

Revert "Add proxy support for fetch_deps."

This reverts commit b6ddb26024118e4eb1413700d0416c836824d710.

Conflicts:
	package.json